% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_over_time_period.R
\name{adjust_over_time_period}
\alias{adjust_over_time_period}
\title{adjust_over_time_period}
\usage{
adjust_over_time_period(x, start, end, baro_file_name = NULL)
}
\arguments{
\item{x}{data.table with values column to adjust}

\item{start}{of period to use for adjustment}

\item{end}{of period to use for adjustment}

\item{id}{optional id to use for adjustment (baro, top transducer, etc.)}
}
\value{
data.table with adjusted values
}
\description{
adjust_over_time_period
}
