% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsk.R
\name{get_rsk_data}
\alias{get_rsk_data}
\title{get_rsk_data}
\usage{
get_rsk_data(rsk)
}
\arguments{
\item{rsk}{the rsk collection}
}
\value{
data.table of rsk results
}
\description{
read an *.rsk file or collection of *.rsk files
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
