% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\docType{data}
\name{LinearRegression}
\alias{LinearRegression}
\alias{rsk_LinearRegression}
\alias{LinearRegression}
\title{Linear Regression}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rsk_LinearRegression

LinearRegression(x, y, fit_intercept = TRUE, normalize = FALSE,
  copy_X = TRUE, n_jobs = 1)
}
\arguments{
\item{x}{matrix. Training Data}

\item{y}{matrix. Target Values}

\item{fit_intercept}{boolean, optional whether to calculate the intercept for
this model. If set to false, no intercept will be used in calculations (e.g.
data is expected to be already centered).}

\item{normalize}{boolean, optional, default False. If True, the regressors X
will be normalized before regression. This parameter is ignored when
fit_intercept is set to False. When the regressors are normalized, note that
this makes the hyperparameters learnt more robust and almost independent of
the number of samples. The same property is not valid for standardized data.
However, if you wish to standardize, please use preprocessing.StandardScaler
before calling fit on an estimator with normalize=False.}

\item{copy_X}{boolean, optional, default True. If True, X will be copied;
else, it may be overwritten.}

\item{n_jobs}{int, optional, default 1. The number of jobs to use for the
computation. If -1 all CPUs are used. This will only provide speedup for}

\item{n_targets}{1 and sufficient large problems.}
}
\description{
Ordinary least squares Linear Regression.
}
\keyword{datasets}
