% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbr_raw_to_temperature}
\alias{rbr_raw_to_temperature}
\title{rbr_raw_to_temperature}
\usage{
rbr_raw_to_temperature(x, calib)
}
\arguments{
\item{calib}{}
}
\value{

}
\description{
rbr_raw_to_temperature
}
