% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbr_temperature_correction}
\alias{rbr_temperature_correction}
\title{rbr_temperature_correction}
\usage{
rbr_temperature_correction(pressure, temperature, x)
}
\arguments{
\item{x}{calibration constants}
}
\value{

}
\description{
rbr_temperature_correction
}
