% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instrument_log.R
\name{read_instrument_log_serial}
\alias{read_instrument_log_serial}
\title{read_instrument_log_serial}
\usage{
read_instrument_log_serial(file_path)
}
\arguments{
\item{file_path}{the path of the log file "ruskin_instrument.log"}
}
\value{
table containing the transducer settings
}
\description{
read_instrument_log_serial
}
