% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev_class.R
\name{read_levelogger}
\alias{read_levelogger}
\title{#' @export
Transducer <- R6Class("Transducer",
               public = list(}
\usage{
read_levelogger(x, transducer_depth = NULL, well_elevation = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
# database name
                 file_name     = NA_character_,
                 file_extension = NA_character_,
}
\details{

}
