% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsk.R
\name{read_rsk}
\alias{read_rsk}
\title{read_rsk}
\usage{
read_rsk(file_name, ...)
}
\arguments{
\item{file_name}{the *rsk file to read}
}
\value{
table with some information about the rsk file
}
\description{
read an *.rsk file or collection of *.rsk files
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
