% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_over_time_period.R
\name{remove_baro}
\alias{remove_baro}
\title{add_baro_column
add a column of barometric pressure to the data.table}
\usage{
remove_baro(x, baro_file_name = NULL, remove_baro = TRUE)
}
\arguments{
\item{x}{data.table with values column to adjust}

\item{baro_file_name}{file name for the baro data}

\item{remove_baro}{remove the baro data from the data.frame}
}
\value{
data.table with adjusted values
}
\description{
add_baro_column
add a column of barometric pressure to the data.table
}
