% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_data.R
\name{rename_data}
\alias{rename_data}
\title{rename_data}
\usage{
rename_data(x)
}
\arguments{
\item{x}{data.table with rbr names to rename}
}
\value{
table with new simplified names
}
\description{
Renames the RBR column heading names
}
