% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk-package.R
\docType{package}
\name{rsk-package}
\alias{rsk}
\alias{rsk-package}
\title{rsk: A tool to work with pressure and temperature datasets from RBR
SQLite3 databases}
\description{
You can learn about the rsk package in the vignettes:
\code{browseVignettes(package = "rsk")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jkennel/rsk}
}

}
\author{
\strong{Maintainer}: Jonathan Kennel \email{jkennel@uoguelph.ca} [translator]

Other contributors:
\itemize{
  \item Beth Parker \email{bparker@uoguelph.ca} [thesis advisor]
}

}
