% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk_read_data_table.R
\name{rsk_read_data_table}
\alias{rsk_read_data_table}
\title{rsk_read_data_table}
\usage{
rsk_read_data_table(db, sql_text)
}
\arguments{
\item{db}{database connection}

\item{sql_text}{character sql string to execute on connection for filtering}
}
\value{
data.table of data from the .rsk file
}
\description{
import sqlite data to R
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
