% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsk_summary.R
\name{rsk_summary}
\alias{rsk_summary}
\title{rsk_summary}
\usage{
rsk_summary(file_name)
}
\arguments{
\item{db}{the database connection}
}
\value{
table with some information about the rsk file
}
\description{
Quick summary of an .rsk file
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
