% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_duckdb.R
\name{to_duckdb}
\alias{to_duckdb}
\title{to_duckdb}
\usage{
to_duckdb(input, output = NULL)
}
\arguments{
\item{input}{the path of the .rsk input file}

\item{output}{the path of the .duckdb output file}
}
\value{
NULL - copies an .rsk file to .duckdb
}
\description{
convert .rsk to .duckdb database.
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
