% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wellcad.R
\name{to_wellcad}
\alias{to_wellcad}
\title{to_wellcad}
\usage{
to_wellcad(input, output = NULL, ...)
}
\arguments{
\item{input}{the path of the .rsk input file}

\item{output}{the path of the .wcd output file}
}
\value{
NULL - export an .rsk file to .wcd (WellCAD)
}
\description{
convert .rsk to .duckdb database
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
