% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsk.R
\name{read_rsk}
\alias{read_rsk}
\title{read_rsk}
\usage{
read_rsk(file_name, return_data_table = FALSE, include_params = NULL, ...)
}
\arguments{
\item{file_name}{the *rsk file to read}

\item{return_data_table}{return a data.table of results?}

\item{include_params}{vector of parameter names contained in *rsk file to
include in returned data.table.}

\item{simplify_names}{do you want the final temperature and pressure results only.
only works when return_data_table is TRUE}
}
\value{
table with some information about the rsk file
}
\description{
read an *.rsk file or collection of *.rsk files
}
\author{
Jonathan Kennel \email{jkennel@uoguelph.ca}
}
